Set NODE_ENV in Dockerfile to development
Out of the box, hackathon-starter is ready for development, hackathons, etc.  There are quite a few post-hackathon or pre-production additions that need to be marked off the list in the case of each application before it is fully ready for a mainstream production environment.

For example, 
- The the login/console.log needs to be redirected to a logging system using Winston or a similar module.  
- Clickstream logging needs to be added.
- Error handling middleware needs to be disabled for production.
- API example routes need to be removed, etc.
- Some assets may need to be moved or called from CDNs.
- Integration (if needed) with a proxy protection and rate limiting services such as CloudFlare.
- Internal or hosting service based SSL support
- etc.

This change is removing the production flag from the Dockerfile to avoid any confusion that the project is ready for full blown production deployment without any changes.