Rework .btn-raised to be custom BS button style. Fixes #1016
This moves definition of the `.btn-raised` out of the `default` theme
definition into the main style project file, so it can be part of the
project style no matter what theme is applied to the project.
If users want to opt-out from that style, it's as easy now as replacing
the `btn-raised` type with e.g built-in one, like `btn-primary`.
The default state of the button tries to retain as much as possible
representation of the button under the default theme definition, whilc
also providing BS based variants for other states of the button like
active, focused, disabled, etc.

Thanks!