Limit husky to non-prod & Upgrade/Fix Docker
Upgraded the docker-compose to version 3.8.
Upgrade the docker Node.js image to 20-slim.
Set mongo image to be the latest instead of a preset version that can go out of date often.
Added the ability to set the docker environment as development or production.
Added a check to the package.json prepare script to only install husky’s execution in non-production environments. This prevents npm install from failing in production environments, where dev dependencies may not be installed. Husky has been a dev dependency.
Fixed a bug where docker-compose was overwriting the application/build folder causing removal of css files that were built with sass.
Removed the extra end of line spaces in docker-compose.yml.